Important : the CodeElement classes should never have a dependency on the generated parser. We want to maintain a clean separation between the grammar and the code object model. The code element builders must centralize the object tree construction logic. I extracted all code element contructors which referenced the generated parser to a SyntaxElementBuilder class. Because the grammar was updated, I had not enough time to fix everything, I had to comment out some parts of these factory methods.
